src/pyproject.toml
[tool.poetry]
name = "processamento-imagem"
version = "1.0.0"
description = "API REST para aplicar filtros de processamento de imagem com diferentes níveis de intensidade"
authors = ["Luiz Felipe, Vincius Vieira, Caio Pierre, Ramiro e Gabriel Soares"]
readme = "../README.md"
packages = [{include = "filtros"}, {include = "modelos"}]

[tool.poetry.dependencies]
python = "^3.12.5"
fastapi = "0.109.0"
uvicorn = {extras = ["standard"], version = "0.27.0"}
python-multipart = "0.0.6"
opencv-python = "4.9.0.80"
scikit-image = "0.22.0"
numpy = "1.26.3"
Pillow = "10.2.0"

[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.4"

[tool.poetry.scripts]
api = "principal:app"

[tool.poe.tasks]
dev = "uvicorn principal:app --reload --host 0.0.0.0 --port 8000"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
